+++ lib/spinor_noise.cu
@@ -207,7 +207,7 @@ namespace quda {
     }
   }
 
-  void spinorNoise(ColorSpinorField &src, int seed, QudaNoiseType type)
+  void spinorNoise(ColorSpinorField &src, unsigned long long seed, QudaNoiseType type)
   {
     RNG *randstates = new RNG(src, seed);
     randstates->Init();
